+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+*   Demo showing controls and dialogs by starting main class
+    `de.muspellheim.commons.fx.DemoApp`.
+
 ### Fixed
 
 *   Date interval picker value was not initialized before control is shown.